Mv torch_mesmer/torch-mesmer/.pys one level up.
